requirements.txt
streamlit==1.32.2
tensorflow==2.12.0
keras==2.12.0
numpy==1.23.5
pillow==10.2.0
h5py==3.8.0